model-containerize/requirements.txt
db-dtypes==1.4.3
google-cloud-bigquery==3.34.0
joblib==1.5.1
numpy==2.1.3
pandas==2.3.0
scikit-learn==1.7.0
scipy==1.15.3
statsmodels==0.14.4
